restructure pages - add note for capture elem in macrowizard page
