+++ mobile/pages/Home/addmacrowizard.page.ts
@@ -1,0 +1,24 @@
+//trigger icon
+//android.widget.LinearLayout[@resource-id="com.arlosoft.macrodroid:id/icons_layout"]/android.widget.ImageView
+//new UiSelector().className("android.widget.ImageView").instance(2)
+//action icon
+//(//android.widget.LinearLayout[@resource-id="com.arlosoft.macrodroid:id/icons_layout"])[2]/android.widget.ImageView
+//new UiSelector().className("android.widget.ImageView").instance(3)
+//constraint icon
+//(//android.widget.LinearLayout[@resource-id="com.arlosoft.macrodroid:id/icons_layout"])[3]/android.widget.ImageView
+//new UiSelector().className("android.widget.ImageView").instance(4)
+////android.widget.ImageButton[@content-desc="Accept"]
+
+//MarcoName and category
+//new UiSelector().resourceId("com.arlosoft.macrodroid:id/enter_name_and_category_spinner")
+//android.widget.Spinner[@resource-id="com.arlosoft.macrodroid:id/enter_name_and_category_spinner"]
+//new UiSelector().resourceId("com.arlosoft.macrodroid:id/enter_name_and_category_name")
+////android.widget.EditText[@resource-id="com.arlosoft.macrodroid:id/enter_name_and_category_name"]
+//handle skip ad here too
+//navigate back to Macro - check data exit
+//add 5 element to Macro page
+//android.widget.TextView[@resource-id="com.arlosoft.macrodroid:id/group_title" and @Text="Sport (1)"] - capture text = Sport (1)
+//android.widget.TextView[@resource-id="com.arlosoft.macrodroid:id/macroNameText" and @Text="Fbb"] - macro test name
+//android.widget.TextView[@resource-id="com.arlosoft.macrodroid:id/macroTrigger" and @Text="Application Removed (Any Application)"]
+//android.widget.TextView[@resource-id="com.arlosoft.macrodroid:id/macroActions" and @Text="Clear Log"]
+//android.widget.TextView[@resource-id="com.arlosoft.macrodroid:id/macroConstraints" and @Text="Airplane Mode Enabled"]
