+++ mobile/pages/Macro/addtrigger.page.ts
@@ -1,4 +1,4 @@
-import BasePage from "./base.page";
+import BasePage from "../base.page";
 class AddTriggerPage extends BasePage {
   async addTriggerType(
     typeTrigger: string,
