+++ mobile/pages/Macro/macro.page.ts
@@ -1,4 +1,4 @@
-import BasePage from "./base.page";
+import BasePage from "../base.page";
 
 class MarcroPage extends BasePage {
   get appNameHeader() {
